requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.0
pandas>=2.3.0
plotly>=6.0.0
sqlalchemy>=2.0.40
lxml>=5.3.0
pyyaml>=6.0
mpxj>=14.0.0
jpype1>=1.5.0
anthropic>=0.45.0
